Modify : 파일 컴포넌트화
