+++ src/components/SubOptBox/SubOptBox.style.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const SubOptBox = styled.div``;
+
+export const NieunSign = styled.div``;
+
+export const SubOptBoxInput = styled.div``;
+
+export const SubOptBoxBtn = styled.div``;
